+++ src/classes/incubation/incubator.ts
@@ -6,11 +6,17 @@
  * license: MIT
  */
 
+// pjson
+import { createRequire } from 'node:module'
+const require = createRequire(import.meta.url)
+const pjson = require('../../../package.json')
+
 // partition
 import yaml from 'js-yaml'
 import fs from 'node:fs'
 import path from 'node:path'
 import shx from 'shelljs'
+import { exec } from '../../lib/utils.js'
 
 import { ICalamaresPartition } from '../../interfaces/i-calamares-partition.js'
 import { IInstaller } from '../../interfaces/i-installer.js'
@@ -30,6 +36,7 @@ const __dirname = path.dirname(new URL(import.meta.url).pathname)
 
 // const branding = require('./branding.js').branding
 import { branding } from './branding.js'
+import Fisherman from './fisherman.js'
 
 /**
  *
@@ -236,9 +243,38 @@ export default class Incubator {
     }
 
     if (Pacman.calamaresExists()) {
-      partitionCustomize()
+        await partitionCustomize()
+        await this.compact()
     }
   }
+
+    /**
+   * 
+   */
+  private async compact() {
+      // modules
+      let path = '/etc/calamares/modules/'
+      const elements = fs.readdirSync(path)
+      for (const elem of elements) {
+        let file = path + elem
+        let fileContent = fs.readFileSync(file, 'utf8')
+        let yamlContent = yaml.load(fileContent)
+        let destContent = `# ${elem}, created by penguins-eggs ${pjson.version}\n`
+        destContent += '---\n'
+        destContent += yaml.dump(yamlContent)
+        fs.writeFileSync(file, destContent, 'utf8')
+      }
+
+      // settings
+      let file='/etc/calamares/settings.conf'
+      let fileContent = fs.readFileSync(file, 'utf8')
+      let yamlContent = yaml.load(fileContent)
+      let destContent = `# settings.conf, created by penguins-eggs ${pjson.version}\n`
+      destContent += '---\n'
+      destContent += yaml.dump(yamlContent)
+      fs.writeFileSync(file, destContent, 'utf8')
+    }
+  
 
   /**
    *
@@ -398,11 +434,15 @@ function write(file: string, content: string, verbose = false) {
 /**
  *
  */
-function partitionCustomize() {
+async function partitionCustomize() {
   const filePartition = '/etc/calamares/modules/partition.conf'
   const partition = yaml.load(fs.readFileSync(filePartition, 'utf8')) as ICalamaresPartition
-  partition.defaultFileSystemType = 'ext4'
+
+  // detect filesystem type
+  let test = await exec(`df -T / | awk 'NR==2 {print $2}'`,{capture: true, echo: false})
+  partition.defaultFileSystemType = test.data.trim()
   partition.availableFileSystemTypes = ['ext4']
+
   if (Pacman.packageIsInstalled('btrfs-progs')) {
     partition.availableFileSystemTypes.push('btrfs')
   }