+++ CHANGELOG.md
@@ -40,6 +40,10 @@ Since version `9.6.x` Penguins' eggs is released - as Debian package - for: `amd
 # Changelog
 Versions are listed on reverse order, the first is the last one. Old versions are moved to [versions](https://sourceforge.net/projects/penguins-eggs/files/DEBS/versions/). 
 
+## penguins-eggs-10.0.46
+* calamares: reformat all calamares modules on `/etc/calamares/`;
+* update to `calamares-3.3.10 for arch.
+
 ## penguins-eggs-10.0.45
 * `eggs love` now not fail if eggs is not configured yet and, configuration is made with `eggs dad --defailt`;
 * `eggs produce` bugfix, `eggs produce` under certain conditions asked for confirmation twice;